requirements.txt
requests==2.31.0
praw==7.7.1
feedparser==6.0.10
airtable-python-wrapper==0.15.3
pyyaml==6.0.1
python-dotenv==1.0.0
schedule==1.2.0
beautifulsoup4==4.12.2
python-dateutil==2.8.2
